[FEAT] : Search 뷰모델 수정, type 추가
